[Chore] CircularBox 컴포넌트 분리
